Add option to rename & approve channel requests
